LOOPCARE-1269: Update BMI calculation and validation

See merge request mobidev/loopcare-frontend!288

+++ lib/core/domain/account/account.dart
@@ -24,7 +24,7 @@ abstract class Account implements _$Account {
     @Default(null) DateTime? groupingStartedAt,
     @Default(0) double height,
     @Default(0) double weight,
-    @Default(0) int bmi,
+    @Default(0) double bmi,
     DateTime? birthDate,
     @Default('') String diabetes,
     @Default('') String? nickname,
